fix issue 46 > created EventSummary and extracted whole component + children to their own file
