+++ src/features/calendar_views/month_view/WeekRow.js
@@ -1,4 +1,4 @@
-import React, { useState, useRef } from "react";
+import React, { useState, useRef, useEffect } from "react";
 
 // PACKAGES
 import PropTypes from "prop-types";
@@ -7,11 +7,11 @@ import dayjs from "dayjs";
 
 // SCRIPTS
 import { sortEvents } from "../../../common/calendar";
-import { range } from "../../../common/utils";
 
 // COMPONENTS
 import Modal from "../../modal/Modal";
 import DayModal from "../day_view/DayModal";
+import EventSummary from "./EventSummary";
 
 function WeekRow({ eventsPerDay, week, month, maxNumberOfEvents }) {
   return (
@@ -125,24 +125,13 @@ RemainingEventsNumber.propTypes = exact({
 function DailyEventList({ date, events, maxNumberOfEvents }) {
   const sortedEvents = sortEvents(events);
 
-  const weekDay = dayjs(date).day();
-  const gridPosition = weekDay === 0 ? 7 : weekDay;
-
-  const eventStyle = `col-start-${gridPosition} mx-2 py-1 px-2 border rounded truncate cursor-pointer z-10 hover:bg-orange-300 hover:border-orange-400`;
-
   return (
     <>
       <div className="contents text-xs">
         {sortedEvents &&
           sortedEvents
             .slice(0, maxNumberOfEvents)
-            .map((event) =>
-              event.allDay ? (
-                <AllDayEvent date={date} event={event} key={event.id} display={event.displayed} css={eventStyle} />
-              ) : (
-                <TimedEvent date={date} event={event} key={event.id} css={eventStyle} />
-              )
-            )}
+            .map((event) => <EventSummary date={date} event={event} key={event.id} />)}
       </div>
     </>
   );
@@ -151,118 +140,6 @@ DailyEventList.propTypes = exact({
   date: PropTypes.string.isRequired,
   events: PropTypes.array,
   maxNumberOfEvents: PropTypes.number.isRequired,
-});
-
-function TimedEvent({ date, event, css }) {
-  const [showModal, setShowModal] = useState(false);
-  const closeModal = () => setShowModal(false);
-  const [chosenEventId, setChosenEventId] = useState(undefined);
-
-  const chooseStyle = () => {
-    const today = dayjs().startOf("day");
-    const currentDate = dayjs(date).startOf("day");
-    if (currentDate.isSame(today)) return "today";
-    return "future";
-  };
-
-  const eventStyle = { today: `border-blue-300 bg-blue-200` };
-  const currentColors = eventStyle[chooseStyle()];
-  const handleClick = () => {
-    setShowModal(true);
-    setChosenEventId(event.id);
-  };
-
-  TimedEvent.propTypes = exact({
-    date: PropTypes.string.isRequired,
-    event: PropTypes.object.isRequired,
-    css: PropTypes.string.isRequired,
-  });
-
-  return (
-    <>
-      <div className={`${css} col-span-1 border-gray-400 ${currentColors} `} key={event.id} onClick={handleClick}>
-        {dayjs(event.start.datetime).format("HH:mm")} {event.title}
-      </div>
-      <Modal showModal={showModal} onCloseModal={closeModal}>
-        <DayModal date={date} chosenEventId={chosenEventId} />
-      </Modal>
-    </>
-  );
-}
-
-function AllDayEvent({ date, event, display, css }) {
-  const remainingDuration = event.duration - event.position;
-
-  const weekDay = dayjs(date).day();
-  const currentWeekDayIndex = weekDay === 0 ? 7 : weekDay;
-
-  const remainingDaysUntilEndOfWeek =
-    currentWeekDayIndex + remainingDuration > 7 ? 7 - currentWeekDayIndex + 1 : remainingDuration + 1;
-
-  const weeklyRemainingDays = range(0, remainingDaysUntilEndOfWeek - 1).map((daysNumber) => {
-    const newDate = dayjs(date, "YYYY-MM-DD").add(daysNumber, "day");
-    return {
-      date: newDate.format("YYYY-MM-DD"),
-      position: event.position + daysNumber,
-    };
-  });
-
-  return (
-    <>
-      {display && (
-        <>
-          <div
-            className={`${css} col-span-${remainingDaysUntilEndOfWeek} bg-indigo-100 border-indigo-200 relative`}
-            key={event.id}
-          >
-            <div className={`absolute gap-2 top-0 left-0  grid grid-cols-${remainingDaysUntilEndOfWeek} w-full h-full`}>
-              {weeklyRemainingDays.map((day) => (
-                <AllDayEventPart day={day} event={event} key={day.date} />
-              ))}
-            </div>
-            {event.title}
-          </div>
-        </>
-      )}
-    </>
-  );
-}
-
-AllDayEvent.propTypes = exact({
-  date: PropTypes.string.isRequired,
-  event: PropTypes.object.isRequired,
-  display: PropTypes.bool.isRequired,
-  css: PropTypes.string.isRequired,
-});
-
-function AllDayEventPart({ day, event }) {
-  const [showModal, setShowModal] = useState(false);
-  const closeModal = () => setShowModal(false);
-  const [chosenEventId, setChosenEventId] = useState(undefined);
-
-  const handleClick = () => {
-    setShowModal(true);
-    setChosenEventId(event.id);
-  };
-
-  return (
-    <>
-      <div
-        className={`border-b-4 border-transparent hover:border-orange-400 first:mx-2 last:mx-2`}
-        onClick={() => handleClick()}
-      >
-        &nbsp;
-      </div>
-      <Modal showModal={showModal} onCloseModal={closeModal}>
-        <DayModal date={day.date} chosenEventId={chosenEventId} />
-      </Modal>
-    </>
-  );
-}
-
-AllDayEventPart.propTypes = exact({
-  day: PropTypes.object.isRequired,
-  event: PropTypes.object.isRequired,
 });
 
 export default WeekRow;
